requirements.txt
flask
google-auth
google-cloud-bigquery
pandas-gbq
uuid